Refactor stat_utils files and move tests to subfolder
