+++ logs/logs.log
@@ -5900,3 +5900,4 @@ Free</div>
 [2022-05-27 15:58:30,706] INFO Scraping surface finished.
 [2022-05-27 15:58:31,708] INFO Generating surface CSV file...
 [2022-05-27 15:58:32,545] INFO CSV file is generated
+[2022-05-30 13:57:12,000] INFO Scraping surface started...
